Modify/ add functions
Modified all the "set" functions to .removeEventListener from rock, paper and scissors variables.
This ensures that the user will not be able to "spam" the buttons. (Planning on adding transitions and animations)

Added a new function checkGameWinner. Checks if computer or player won 3 times.
If there is a winner then another function is called which will either be a "Game Winner" screen or "Game Over" screen.
However, if there is no game winner, then .addEventListener is attached to the variables rock, paper and scissors again.