Add try again options
Added two separate functions that adds a "show" class to an HTML element depending on
who wins 5 times first (computer or player). When "show" class is added to element, a modal
box will appear on the webpage and the user will have an option to "restart game" by clicking
a button. If button is clicked, then scores are zeroed out and "show" class is removed from
HTML elements. In addition, .addEventListeners will be added back to rock, paper and scissors
variables.